Refactor code
Add username package
