+++ src/main/java/ru/mail/polis/DAOFactory.java
@@ -17,6 +17,7 @@
 package ru.mail.polis;
 
 import org.jetbrains.annotations.NotNull;
+import ru.mail.polis.vladislav.DAOImpl;
 
 import java.io.File;
 import java.io.IOException;
