+++ src/main/java/ru/mail/polis/vladislav/DAOImpl.java
@@ -1,7 +1,9 @@
-package ru.mail.polis;
+package ru.mail.polis.vladislav;
 
 import com.google.common.collect.Iterators;
 import org.jetbrains.annotations.NotNull;
+import ru.mail.polis.DAO;
+import ru.mail.polis.Record;
 
 import java.nio.ByteBuffer;
 import java.util.Iterator;
